requirements.txt
# =============================================================================
# Python Dependencies — Azure Databricks Lakehouse
# =============================================================================

# Core
pyspark==3.5.1
delta-spark==3.1.0
pyyaml>=6.0

# Data Quality
great-expectations>=0.18.0

# Testing
pytest>=8.0
pytest-cov>=5.0
chispa>=0.10.1          # PySpark DataFrame testing utilities

# Code Quality
ruff>=0.3.0
black>=24.0

# Databricks
databricks-sdk>=0.20.0

# Infrastructure
azure-storage-file-datalake>=12.14.0
azure-identity>=1.15.0
